requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.66.1
beautifulsoup4>=4.12.2
aiohttp>=3.9.1
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1